Hacked the ADSR envelope to work at least a little bit better. Instead of recording a current amplitude in a float, it now uses a double, so long envelopes are no longer too short (unless they're really long).

+++ Modules/Adsr.hpp
@@ -72,9 +72,9 @@ class Adsr: public Module
 {
     public:
         // The current amplitudes
-        float current_amplitude;
+        double current_amplitude;
         // Which phase of the envelope this module is in
-        float phase_num;
+        int phase_num;
 
         // Constructor and destructor
         Adsr();
